Centralize jax_enable_x64 config; move chunk.py into backends/
registry.py, compiler.py and statevector.py each toggled
jax_enable_x64 unconditionally at module import time, so merely
`import dense_evolution` silently overrode a precision a caller had
already configured for unrelated JAX code in the same process --
mps.py's own docstring already documents the opposite convention
(enable lazily, let the caller decide). dense_evolution/config.py
centralizes this: ensure_x64() is called lazily by the specific
callers that need complex128 (DenseSVSimulator, circuit_to_energy_fn),
and set_precision() is the public opt-out. tools/dashboard/core sets
precision explicitly now, since some of its paths (MPSSimulator-only)
never construct a DenseSVSimulator and previously relied on the
removed import-time side effect.

chunk.py was the one module left at the package root when the rest of
dense_evolution was split into subpackages (simulator, compiler,
gates, trotter, qec, ...). Moved it to dense_evolution/backends/chunk.py
with the rest of the execution backends; the old import path keeps
working via a shim that aliases dense_evolution.chunk directly onto
the real module object in sys.modules, since tests and dashboard code
import private helpers from it by module path, not just the public
Chunk class.

Full test suite: 943 passed, 19 skipped, 0 failed.
